Update minq encoded files
This is after changing the encoder to not send table updates when
there is too much outstanding.  This regresses performance in a few
case but is generally positive.

The "noack" scenarios all improve.  No effort is wasted on inserts that
won't be used.

The "ack" scenarios sometimes improve for the same reason, but several
get worse.  In that case, I surmise that the acknowledgments allow the
table to be refreshed with more relevant information for subsequent
requests and refusing to use the table in this way turns out to be
suboptimal for values that do end up being repeated.

The net gain is almost 1Mb across this set, so I'm keeping the change,
even if this setup says virtually nothing about real performance.